Renamed module from stitch to realm
