+++ package.json
@@ -2,7 +2,7 @@
   "name": "stitch-context-module",
   "version": "1.0.0",
   "description": "A local stitch context that can be used to test MongoDB Stitch Functions locally",
-  "main": "stitch-context.js",
+  "main": "realm-context.js",
   "scripts": {
     "test": "mocha tests/*"
   },
